requirements1.txt
streamlit
requests
json
ast
tempfile
os
datetime 
langchain
langchain-community
langchain-core
langchain-huggingface
langchain-text-splitters
langchain_community.embeddings 
langdetect
langsmith
chromadb
httpx
pypdf
msoffcrypto-tool
docx
docx2txt
html5lib
jq
lxml
openpyxl
python-docx
unstructured